change the typr of variable cont

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -6,7 +6,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int cont;
+	size_t cont;
 
 	while (h != NULL)
 	{
